BREAKING CHANGE: add rabbitmq
