--- services/ocr/send-test.py
@@ -1,17 +1,0 @@
-
-import pika
-
-#RABBITMQ_HOST = "rabbitmq"
-RABBITMQ_HOST = "localhost"
-
-connection = pika.BlockingConnection(pika.ConnectionParameters(RABBITMQ_HOST))
-channel = connection.channel()
-
-channel.queue_declare(queue='hello')
-
-channel.basic_publish(exchange='',
-                      routing_key='hello',
-                      body='Hello World!')
-print(" [x] Sent 'Hello World!'")
-
-connection.close()
